Refactor project structure and remove deprecated files
- Updated .gitignore to exclude pytest cache and output directory.
- Deleted unused scripts: intervals_calendar.py, monitor.py, peloton_matcher.py, plan_generator.py, reasoner.py, strava_api.py, tour_coach.py, and readiness_score_log.csv.
- Moved relevant code to src/lanterne_rouge directory for better organization.
- Implemented new functionality in the monitor and plan_generator modules.
- Added tests for plan generation and reasoning logic.
- Updated requirements.txt to include pytest for testing.
