--- readiness_score_log.csv
@@ -1,6 +1,0 @@
-day,readiness_score,hrv_balance
-2025-04-27,84,78
-2025-04-27,84,78
-2025-04-27,84,78
-2025-04-27,84,78
-2025-04-27,84,78
