+++ requirements.txt
@@ -1,2 +1,3 @@
-requests
-python-dotenv
+python-dotenv>=1.0.0
+requests>=2.28.0
+pytest>=8.0.0
