+++ src/lanterne_rouge/archive/intervals_calendar.py
