+++ src/lanterne_rouge/peloton_matcher.py
@@ -1,3 +1,4 @@
+
 # peloton-matcher.py
 
 def match_peloton_class(workout_type):
