+++ src/lanterne_rouge/plan_generator.py
