+++ src/lanterne_rouge/reasoner.py
