--- tour_coach_update.txt
@@ -1,22 +1,0 @@
-Date: Sunday, April 27, 2025
-
-Planned Workout:
-- Long Endurance Ride (2hr @ Zone 2–3 with surges @ Zone 4)
-
-Readiness and Recovery:
-- Readiness Score: 84
-- Readiness Day: 2025-04-27
-- CTL (Fitness): 31.191942905006744
-- ATL (Fatigue): 50.20563189757815
-- TSB (Form): -19.013688992571407
-
-Recommendation:
-- ⚠️ Form is moderately negative (TSB < -10). Training OK but monitor fatigue.
-
-Peloton Class Suggestion:
-- 90 min Power Zone Endurance Ride with Matt Wilpers
-
-Notes:
-- Adjust fueling based on today's effort. Stay hydrated and recover smart.
-
-Generated by Tour Coach Agent v0.2.0
